moved move desk functions to hardware_controller
